squash util in exchange
