+++ wayfinder_paths/adapters/hyperliquid_adapter/exchange.py
@@ -24,11 +24,6 @@ from hyperliquid.utils.types import BuilderInfo
 from loguru import logger
 from web3 import Web3
 
-from wayfinder_paths.adapters.hyperliquid_adapter.util import (
-    get_price_decimals_for_hypecore_asset,
-    sig_hex_to_hl_signature,
-)
-
 ARBITRUM_CHAIN_ID = "0xa4b1"
 MAINNET = "Mainnet"
 USER_DECLINED_ERROR = {
@@ -48,6 +43,30 @@ class Exchange:
         self.api = API()
         self.sign_callback = sign_callback
         self.signing_type = signing_type
+
+    def get_sz_decimals(self, asset_id: int) -> int:
+        return self.info.asset_to_sz_decimals[asset_id]
+
+    def get_price_decimals(self, asset_id: int) -> int:
+        is_spot = asset_id >= 10_000
+        return (6 if not is_spot else 8) - self.get_sz_decimals(asset_id)
+
+    @staticmethod
+    def sig_hex_to_hl_signature(sig_hex: str) -> dict[str, Any]:
+        """Convert a 65-byte hex signature into Hyperliquid {r,s,v}."""
+        if not isinstance(sig_hex, str) or not sig_hex.startswith("0x"):
+            raise ValueError("Expected hex signature string starting with 0x")
+        raw = bytes.fromhex(sig_hex[2:])
+        if len(raw) != 65:
+            raise ValueError(f"Expected 65-byte signature, got {len(raw)} bytes")
+
+        r = raw[0:32]
+        s = raw[32:64]
+        v = raw[64]
+        if v < 27:
+            v += 27
+
+        return {"r": f"0x{r.hex()}", "s": f"0x{s.hex()}", "v": int(v)}
 
     def _create_hypecore_order_actions(
         self,
@@ -93,7 +112,7 @@ class Exchange:
         price = midprice * ((1 + slippage) if is_buy else (1 - slippage))
         price = round(
             float(f"{price:.5g}"),
-            get_price_decimals_for_hypecore_asset(self.info, asset_id),
+            self.get_price_decimals(asset_id),
         )
         order_actions = self._create_hypecore_order_actions(
             asset_id,
@@ -291,7 +310,7 @@ class Exchange:
             sig_hex = await self.sign_callback(payload)
             if not sig_hex:
                 return None
-            return sig_hex_to_hl_signature(sig_hex)
+            return self.sig_hex_to_hl_signature(sig_hex)
 
         payload = encode_typed_data(full_message=payload)
         return await self.sign_callback(action, payload, address)
